requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
flask>=3.0.0
pillow>=10.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
